Three exceptions moved to importhelper package
