+++ web-ui/src/main/java/org/archcnl/domain/common/io/importhelper/PredicateParser.java
@@ -7,7 +7,7 @@ import org.archcnl.domain.common.conceptsandrelations.CustomRelation;
 import org.archcnl.domain.common.conceptsandrelations.Relation;
 import org.archcnl.domain.common.io.RegexUtils;
 import org.archcnl.domain.common.io.exceptions.NoMatchFoundException;
-import org.archcnl.domain.input.exceptions.NoRelationException;
+import org.archcnl.domain.common.io.importhelper.exceptions.NoRelationException;
 
 public class PredicateParser {
