+++ web-ui/src/main/java/org/archcnl/domain/common/io/importhelper/QueryParser.java
@@ -15,7 +15,7 @@ import org.archcnl.domain.common.conceptsandrelations.andtriplets.triplet.Variab
 import org.archcnl.domain.common.conceptsandrelations.andtriplets.triplet.exceptions.InvalidVariableNameException;
 import org.archcnl.domain.common.io.RegexUtils;
 import org.archcnl.domain.common.io.exceptions.NoMatchFoundException;
-import org.archcnl.domain.input.exceptions.NoTripletException;
+import org.archcnl.domain.common.io.importhelper.exceptions.NoTripletException;
 import org.archcnl.domain.output.model.query.FreeTextQuery;
 import org.archcnl.domain.output.model.query.Query;
 import org.archcnl.domain.output.model.query.SelectClause;
