+++ web-ui/src/main/java/org/archcnl/domain/common/io/importhelper/exceptions/NoObjectTypeException.java
@@ -1,4 +1,4 @@
-package org.archcnl.domain.input.exceptions;
+package org.archcnl.domain.common.io.importhelper.exceptions;
 
 public class NoObjectTypeException extends Exception {
