+++ web-ui/src/main/java/org/archcnl/domain/common/io/importhelper/exceptions/NoRelationException.java
@@ -1,4 +1,4 @@
-package org.archcnl.domain.input.exceptions;
+package org.archcnl.domain.common.io.importhelper.exceptions;
 
 public class NoRelationException extends Exception {
