+++ web-ui/src/main/java/org/archcnl/domain/common/io/importhelper/exceptions/NoTripletException.java
@@ -1,4 +1,4 @@
-package org.archcnl.domain.input.exceptions;
+package org.archcnl.domain.common.io.importhelper.exceptions;
 
 public class NoTripletException extends Exception {
